[C File struct]: refacto
