+++ lib/cbindings/display/display.c
@@ -13,6 +13,8 @@
 #include <sys/ioctl.h>
 #include <termios.h>
 #include <string.h>
+#include "termove.h"
+#include "display.h"
 
 #define RGBA "RGBA"
 #define NCHANNEL 4
@@ -20,16 +22,6 @@
 
 #define NO_FILE "The list is empty"
 #define NO_IMAGE_FILE "Not an image file"
-
-const char* NEW_SCREEN_BUFF_SEQ = "\033[?1049h\033[H";
-const char* END_SRCEEN_BUFF_SEQ = "\033[?1049l";
-const char* CLEAR_CONSOLE = "\033[2J";
-const char* UPPER_LEFT_CORNER = "┌";
-const char* UPPER_RIGHT_CORNER = "┐";
-const char* LOWER_LEFT_CORNER = "└";
-const char* LOWER_RIGTH_CORNER = "┘";
-const char* HORIZONTAL_LINE = "─"; // "─" != '-'
-const char* VERTICAL_LINE = "│";
 
 struct termios raw;
 struct termios orig_termios;
@@ -51,36 +43,39 @@ void disableRawMode() {
   tcsetattr(STDIN_FILENO, TCSAFLUSH, &orig_termios);
 }
 
+void start_window() {
+    enableRawMode();
+    write(STDOUT_FILENO, NEW_SCREEN_BUFF_SEQ, strlen(NEW_SCREEN_BUFF_SEQ));
+    
+}
+
+void end_window() {
+    write(STDOUT_FILENO, END_SRCEEN_BUFF_SEQ, strlen(END_SRCEEN_BUFF_SEQ));
+    disableRawMode();
+}
+
 void handle_sigint(int signo) {
     disableRawMode();
     end_window();
 }
 
+image_array_t* create_image_array(size_t nbimage) {
+    image_array_t* array = malloc(sizeof(image_array_t));
+    if (!array) return NULL;
 
-typedef enum pixel_mode {
-    ITERM = 0,
-    KITTY,
-    SIXEL,
-    NONE
-} pixel_mode_t;
+    image_t* images_ptr = malloc(sizeof(image_t) * nbimage);
+    if (!images_ptr) return NULL;
 
-typedef enum {
-    NO_ERROR,
-    MALLOC_FAIL,
-    MAGICKNULL,
-    MAGICK_EXPORT_FAIl,
-} exit_status_t;
+    array->images = images_ptr;
+    array->count = nbimage;
+    return array;
+}
 
-typedef struct {
-    const char* image_name;
-    const unsigned char* image_bytes;
-    size_t bytes_len;
-} image_t;
 
-typedef struct {
-    image_t* images;
-    size_t count;
-} image_array_t;
+void free_image_array(image_array_t* array) {
+    free((void *) array->images);
+    free(array);
+}
 
 void set_pixel_mode(ChafaCanvasConfig* config, pixel_mode_t mode) {
     switch (mode) {
@@ -98,75 +93,7 @@ void set_pixel_mode(ChafaCanvasConfig* config, pixel_mode_t mode) {
     }
 }
 
-void set_cursor_at(unsigned int line, unsigned int colmn) {
-    fprintf(stdout, "\033[%u;%uf", line, colmn);
-    fflush(stdout);
-}
 
-void move_down(unsigned int l) {
-    fprintf(stdout, "\033[%uB", l);
-    fflush(stdout);
-}
-
-void clear(){
-    fprintf(stdout, "%s", CLEAR_CONSOLE);
-    fflush(stdout);
-}
-
-void move_forward_column(unsigned int c) {
-    fprintf(stdout, "\033[%uC", c);
-    fflush(stdout);
-}
-
-
-
-void start_window() {
-    enableRawMode();
-    write(STDOUT_FILENO, NEW_SCREEN_BUFF_SEQ, strlen(NEW_SCREEN_BUFF_SEQ));
-    
-}
-
-void end_window() {
-    write(STDOUT_FILENO, END_SRCEEN_BUFF_SEQ, strlen(END_SRCEEN_BUFF_SEQ));
-    disableRawMode();
-}
-
-void draw_vertical_line() {
-    fprintf(stdout, "%s", VERTICAL_LINE);
-    fflush(stdout);
-}
-
-void draw_gstring(GString* gstring){
-    fwrite (gstring->str, sizeof (char), gstring->len, stdout);
-    fflush(stdout);
-}
-
-void draw_horizontal_line() {
-    fprintf(stdout, "%s", HORIZONTAL_LINE);
-    fflush(stdout);
-}
-
-void draw_string(const char* s) {
-    fprintf(stdout, "%s", s);
-    fflush(stdout);
-}
-
-void draw_char(char c) {
-    fprintf(stdout, "%c", c);
-    fflush(stdout);
-}
-
-void next_line(unsigned int current_line) {
-    set_cursor_at(current_line + 1, 0);
-    fflush(stdout);
-}
-
-void redraw_empty(const struct winsize* w){
-    set_cursor_at(0, 0);
-    for (unsigned int i = 0; i < w->ws_col * w->ws_row; i += 1) {
-        draw_char(' ');
-    } 
-}
 
 void draw_first_line(const char* title, const struct winsize *w, int endline) {
     size_t title_len = strlen(title);
@@ -308,24 +235,6 @@ CAMLprim value caml_chafa_test(value path, value unit) {
     MagickWandTerminus();
     free((void *) pixels);
     CAMLreturn(Val_unit);
-}
-
-image_array_t* create_image_array(size_t nbimage) {
-    image_array_t* array = malloc(sizeof(image_array_t));
-    if (!array) return NULL;
-
-    image_t* images_ptr = malloc(sizeof(image_t) * nbimage);
-    if (!images_ptr) return NULL;
-
-    array->images = images_ptr;
-    array->count = nbimage;
-    return array;
-}
-
-
-void free_image_array(image_array_t* array) {
-    free((void *) array->images);
-    free(array);
 }
