+++ lib/cbindings/dune
@@ -19,13 +19,15 @@
 
 (library
  (name cbindings)
-  (foreign_stubs (language c) (names input display)
+  (install_c_headers display termove)
+  (foreign_stubs (language c) (names input display termove)
     (flags 
       (:include cflags_chafa.sexp) 
       (:include cflags_MagickWand.sexp)
       (-fPIC)
     )
   )
+
   (c_library_flags 
     (:include clibrary_flags_chafa.sexp)
     (:include clibrary_flags_MagickWand.sexp)
